перенес функционал определния рейтинаг пользователя в отдельную утилиту
