+++ src/main.js
@@ -1,5 +1,6 @@
 import {InsertPlace} from './const';
 import {render} from './utils/render';
+import {getProfileRank} from './utils/rating';
 
 import UserProfileView from './view/user-profile';
 import FilmsFilterView from './view/films-filter';
@@ -19,28 +20,6 @@ import FilmsTotalView from './view/films-total';
 import {Count as DataCount} from './mock/const-data';
 import {getFilmData} from './mock/film-data';
 import {getCommentData} from './mock/comment-data';
-
-const profileRatings = [
-  {
-    title: 'Novice',
-    range: {
-      min: 1,
-      max: 10,
-    },
-  }, {
-    title: 'Fan',
-    range: {
-      min: 11,
-      max: 20,
-    },
-  }, {
-    title: 'Movie Buff',
-    range: {
-      min: 21,
-      max: Number.MAX_SAFE_INTEGER,
-    },
-  },
-];
 
 const FilmCount = {
   STEP: 5,
@@ -68,10 +47,6 @@ const filterResult = films.reduce((sum, {isWatchlist, isWatched, isFavorite}) =>
 
 const filmsTotalCount = films.length;
 const watchedFilmsCount = filterResult.watchedCount;
-
-const getProfileRank = (value) => {
-  return value !== 0 ? (profileRatings.find((item) => value >= item.range.min && value <= item.range.max)).title : '';
-};
 
 const rank = getProfileRank(watchedFilmsCount);
